Milestone #4A - Ready for passoff

+++ server/src/main/java/edu/byu/cs/tweeter/server/service/FollowService.java
@@ -19,7 +19,6 @@ import edu.byu.cs.tweeter.model.net.response.GetFollowingCountResponse;
 import edu.byu.cs.tweeter.model.net.response.GetFollowingResponse;
 import edu.byu.cs.tweeter.model.net.response.IsFollowerResponse;
 import edu.byu.cs.tweeter.model.net.response.UnfollowResponse;
-import edu.byu.cs.tweeter.server.dao.DAOException;
 import edu.byu.cs.tweeter.server.dao.FollowDAO;
 import edu.byu.cs.tweeter.server.dao.UserDAO;
 import edu.byu.cs.tweeter.server.dao.dynamo.FollowDynamoDAO;
@@ -27,7 +26,7 @@ import edu.byu.cs.tweeter.server.dao.dynamo.FollowDynamoDAO;
 /**
  * Contains the business logic for getting the users a user is following.
  */
-public class FollowService {
+public class FollowService extends Service {
     private final FollowDAO followDAO;
     private final UserDAO userDAO;
 
@@ -48,6 +47,7 @@ public class FollowService {
         return this.followDAO;
     }
 
+    @Override
     public UserDAO getUserDAO() { return this.userDAO; }
 
     /**
@@ -66,22 +66,31 @@ public class FollowService {
             throw new RuntimeException("[BadRequest] Request needs to have a positive limit");
         }
 
+        if (!authenticate(request.getAuthToken())) {
+            return new GetFollowersResponse("Unable to authenticate! Your session may have expired. Please log out and log back in.");
+        }
+
+        System.out.printf("Last alias: %s%n", request.getLastItem() != null ? request.getLastItem().getAlias() : "null");
         List<String> followerAliases;
         try {
             followerAliases = getFollowDAO().getFollowers(request.getTargetUser().getAlias(),
                     request.getLimit(), request.getLastItem() != null ? request.getLastItem().getAlias() : null);
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to get list of followers: " + e.getMessage());
         }
 
-        if (followerAliases == null) {
+        if (followerAliases == null || followerAliases.isEmpty()) {
             return new GetFollowersResponse(new ArrayList<>(), false);
         }
+
+        System.out.println(followerAliases);
 
         try {
             return new GetFollowersResponse(getUserDAO().batchGetUsers(followerAliases),
                     followerAliases.size() == request.getLimit());
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to get users based on followers: " + e.getMessage());
         }
     }
@@ -102,22 +111,28 @@ public class FollowService {
             throw new RuntimeException("[BadRequest] Request needs to have a positive limit");
         }
 
+        if (!authenticate(request.getAuthToken())) {
+            return new GetFollowingResponse("Unable to authenticate! Your session may have expired. Please log out and log back in.");
+        }
+
         List<String> followeeAliases;
         try {
             followeeAliases = getFollowDAO().getFollowees(request.getTargetUser().getAlias(),
                     request.getLimit(), request.getLastItem() != null ? request.getLastItem().getAlias() : null);
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to get list of followees: " + e.getMessage());
         }
 
-        if (followeeAliases == null) {
+        if (followeeAliases == null || followeeAliases.isEmpty()) {
             return new GetFollowingResponse(new ArrayList<>(), false);
         }
 
         try {
             return new GetFollowingResponse(getUserDAO().batchGetUsers(followeeAliases),
                     followeeAliases.size() == request.getLimit());
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to get users based on followees: " + e.getMessage());
         }
     }
@@ -133,21 +148,28 @@ public class FollowService {
             throw new RuntimeException("[BadRequest] Request needs to have an authToken");
         }
 
+        if (!authenticate(request.getAuthToken())) {
+            return new FollowResponse("Unable to authenticate! Your session may have expired. Please log out and log back in.");
+        }
+
         try {
             getFollowDAO().putFollower(request.getFollowee().getAlias(), request.getLoggedInUser().getAlias());
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to add follower: " + e.getMessage());
         }
 
         try {
             getUserDAO().incrementFollowingCount(request.getLoggedInUser().getAlias(), 1);
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to update following count: " + e.getMessage());
         }
 
         try {
             getUserDAO().incrementFollowerCount(request.getFollowee().getAlias(), 1);
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to update followers count for user \"" +
                 request.getFollowee().getAlias() + "\": " + e.getMessage());
         }
@@ -166,21 +188,28 @@ public class FollowService {
             throw new RuntimeException("[BadRequest] Request needs to have an authToken");
         }
 
+        if (!authenticate(request.getAuthToken())) {
+            return new UnfollowResponse("Unable to authenticate! Your session may have expired. Please log out and log back in.");
+        }
+
         try {
             getFollowDAO().deleteFollower(request.getUnfollowee().getAlias(), request.getLoggedInUser().getAlias());
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to delete follower: " + e.getMessage());
         }
 
         try {
             getUserDAO().incrementFollowingCount(request.getLoggedInUser().getAlias(), -1);
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to update following count: " + e.getMessage());
         }
 
         try {
             getUserDAO().incrementFollowerCount(request.getUnfollowee().getAlias(), -1);
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to update followers count for user \"" +
                     request.getUnfollowee().getAlias() + "\": " + e.getMessage());
         }
@@ -199,17 +228,16 @@ public class FollowService {
             throw new RuntimeException("[BadRequest] Request needs to have an authToken");
         }
 
-        try {
-            if (!getUserDAO().authenticate(request.getAuthToken())) {
-                return new IsFollowerResponse("Unable to authenticate. Try logging out and logging back in.");
-            }
-        } catch (DAOException e) {
-            throw new RuntimeException("[DBError] Unable to authenticate user: " + e.getMessage());
+        if (!authenticate(request.getAuthToken())) {
+            return new IsFollowerResponse("Unable to authenticate! Your session may have expired. Please log out and log back in.");
         }
 
         try {
-            return new IsFollowerResponse(getFollowDAO().isFollower(request.getFollowee().getAlias(), request.getFollower().getAlias()));
-        } catch (DAOException e) {
+            boolean isFollower = getFollowDAO().isFollower(request.getFollowee().getAlias(), request.getFollower().getAlias());
+            System.out.println("isFollower is " + isFollower);
+            return new IsFollowerResponse(isFollower);
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to determine follow relationship between \"" +
                 request.getFollowee().getAlias() + "\" and \"" + request.getFollower().getAlias() + "\"");
         }
@@ -223,17 +251,14 @@ public class FollowService {
             throw new RuntimeException("[BadRequest] Request needs to have an authToken");
         }
 
-        try {
-            if (!getUserDAO().authenticate(request.getAuthToken())) {
-                return new GetFollowersCountResponse("Unable to authenticate. Try logging out and logging back in.");
-            }
-        } catch (DAOException e) {
-            throw new RuntimeException("[DBError] Unable to authenticate user: " + e.getMessage());
+        if (!authenticate(request.getAuthToken())) {
+            return new GetFollowersCountResponse("Unable to authenticate! Your session may have expired. Please log out and log back in.");
         }
 
         try {
             return new GetFollowersCountResponse(getUserDAO().getFollowersCount(request.getTargetUser().getAlias()));
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to get follower count: " + e.getMessage());
         }
     }
@@ -246,17 +271,14 @@ public class FollowService {
             throw new RuntimeException("[BadRequest] Request needs to have an authToken");
         }
 
-        try {
-            if (!getUserDAO().authenticate(request.getAuthToken())) {
-                return new GetFollowingCountResponse("Unable to authenticate. Try logging out and logging back in.");
-            }
-        } catch (DAOException e) {
-            throw new RuntimeException("[DBError] Unable to authenticate user: " + e.getMessage());
+        if (!authenticate(request.getAuthToken())) {
+            return new GetFollowingCountResponse("Unable to authenticate! Your session may have expired. Please log out and log back in.");
         }
 
         try {
             return new GetFollowingCountResponse(getUserDAO().getFollowingCount(request.getTargetUser().getAlias()));
-        } catch (DAOException e) {
+        } catch (Exception e) {
+            e.printStackTrace();
             throw new RuntimeException("[DBError] Unable to get followee count: " + e.getMessage());
         }
     }
